pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "moolah-iq-pipeline"
version = "3.8.2"
description = "12-agent LangGraph system for automated personal finance video production"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Moolah IQ" },
]

dependencies = [
    "langgraph",
    "langchain",
    "langchain-anthropic",
    "langchain-google-genai",
    "langsmith",
    "langchain-ollama",
    "httpx",
    "python-dotenv",
    "pyyaml",
    "ffmpeg-python",
    "pillow",
    "requests",
    "aiohttp",
    "websockets",
    "fastapi",
    "uvicorn",
    "faster-whisper",
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
